requirements.txt
asgiref==3.2.4
Django==2.0
django-elasticsearch-dsl==0.5.1
django-elasticsearch-dsl-drf==0.5
django-nine==0.2.3
djangorestframework==3.11.0
elasticsearch==5.5.3
elasticsearch-dsl==5.0.0
ipaddress==1.0.23
mysqlclient==1.4.6
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
